--- a/ggml/src/ggml-vulkan/vulkan-shaders/im2col.comp
+++ b/ggml/src/ggml-vulkan/vulkan-shaders/im2col.comp
@@ -48,7 +48,7 @@ void main() {
 
     [[unroll]] for (uint idx = 0; idx < NUM_ITER; ++idx) {
 
-        uint i = base_idx + idx;
+        const uint i = base_idx + idx;
 
         if (i >= p.pelements) {
             break;
@@ -59,8 +59,8 @@ void main() {
         const uint ky  = rem / p.OW;
         const uint ix  = rem % p.OW;
 
-        int iiw = int(ix) * p.s0 + int(kx) * p.d0 - p.p0;
-        int iih = oh_s1 + int(ky) * p.d1 - p.p1;
+        const int iiw = int(ix) * p.s0 + int(kx) * p.d0 - p.p0;
+        const int iih = oh_s1 + int(ky) * p.d1 - p.p1;
 
         const uint dst_offset = dst_base + ix * p.CHW + ky * p.KW + kx;